feat: NetworkError타입 생성 및 Error타입에 맞춰 Network메서드 데이터 핸들링
